+++ src/main/java/tn/esprit/gestionprojet/services/IEquipeService.java
@@ -8,6 +8,6 @@ public interface IEquipeService {
     public Equipe addEquipe(Equipe equipe);
     public Equipe updateEquipe(Equipe equipe);
     public List<Equipe> getAllEquipes();
-    public Equipe getEquipeById(int id);
-    public void deleteEquipeById(int id);
+    public Equipe getEquipeById(Long id);
+    public void deleteEquipeById(Long id);
 }
